pyproject.toml
[tool.poetry]
name = "pr-reviewer"
package-mode = false
version = "1.0.0"
description = "AI-powered Pull Request Reviewer"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.0.0"
PyGithub = "^2.1.1"
gitpython = "^3.1.40"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 